pyproject.toml
[tool.poetry]
name = "geektask"
version = "0.1.0"
description = ""
authors = ["SolioKyaiz <jonesbonked@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.7"
djangorestframework = "^3.16.0"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
